v0.1.16

+++ common/spatial/flatgeobuf_test.go
@@ -161,7 +161,8 @@ func TestWriteFlatGeobufWritesEWKBFeatures(t *testing.T) {
 			Geometry: ewkbValue,
 			Properties: map[string]interface{}{
 				"name":   "Beijing",
-				"rank":   int64(1),
+				// JSON-backed workflow batches decode integer JSON numbers as float64.
+				"rank":   float64(1),
 				"area":   "43854.25",
 				"active": true,
 			},
